encoding_index_simpchinese

Module gb18030

GB 18030 two-byte area.

From the Encoding Standard:

This matches the GB18030 standard for code points encoded as two bytes, except 0xA3 0xA0 maps to U+3000 to be compatible with deployed content.

Functions

backward

Returns the index pointer for code point code in this index.

forward

Returns the index code point for pointer code in this index.