encoding_index_simpchinese

Module gb18030_ranges

GB 18030 four-byte area.

From the Encoding Standard:

This index works different from all others. Listing all code points would result in over a million items whereas they can be represented neatly in 207 ranges combined with trivial limit checks. It therefore only superficially matches the GB18030 standard for code points encoded as four bytes.

Functions

backward

Returns the index pointer for code point code in this index.

forward

Returns the index code point for pointer code in this index.